feat(payment): Allow changing subscription payment method
Adds functionality for users to select a new payment method for individual subscriptions. This involves updating the payment method UI to accept a `setCurrentView` prop and introducing a modal within the `Subscriptions` component for selecting and updating payment methods.

Also, updates the mock Stripe data to reflect an expired card for testing purposes.